Update docker.md
I added some troubleshooting to the "Build the containers for the first time"-section. I followed the instructions but the sail build failed until I changed all of this! @jankapunkt we need to include this in a better and structured way. maybe it would be good to edit the .env and the docker-compose.yml file before doing the sail build?