feat(teams): Load team data on-demand
Team data is no longer pre-loaded on the project overview page. Instead, it is now fetched from a dedicated API endpoint when the "Collaboration" tab is clicked. This reduces the number of initial database queries and improves page load times.
In addition to that, I removed the "ownTeams" from HandleInertiaRequests.php so we don't load that in every page, but only when it's needed.