GroupedBarChart Visualizer
This visualization shows the code counts grouped for each source. Nice! It was quite complicated to create the data object for plotly. I think it would be better to create a kind for R-like DATAFRAME in the API, so that all visualization plugins can use this, because otherwise you really have to make sure that you don't introduce errors.